GET /api/restaurants/top
