+++ controllers/api/restController.js
@@ -21,7 +21,13 @@ const restController = {
 
   getDashboard: (req, res) => {
     restService.getDashboard(req, res, data => {
-        return res.json(data);
+      return res.json(data);
+    });
+  },
+
+  getTopRestaurant: (req, res) => {
+    restService.getTopRestaurant(req, res, data => {
+      return res.json(data);
     });
   }
 };
