+++ routes/apis.js
@@ -29,6 +29,7 @@ router.get(
 );
 router.get('/restaurants', authenticated, restController.getRestaurants);
 router.get('/restaurants/feeds', authenticated, restController.getFeeds);
+router.get('/restaurants/top', authenticated, restController.getTopRestaurant);
 router.get('/restaurants/:id', authenticated, restController.getRestaurant);
 router.get(
   '/restaurants/:id/dashboard',
